refactor: layout, providers 디렉토리 분리
