+++ apps/frontend/src/_components/layout/providers/RootProviders.tsx
@@ -1,18 +1,14 @@
-import Layout from "../_layout";
-
 import ReduxStoreProvider from "./ReduxStoreProvider";
 import QueryProviders from "./QueryProvider";
 
-export default function LayoutWithProviders({
+export default function RootProviders({
   children,
 }: {
   children: React.ReactNode;
 }) {
   return (
     <QueryProviders>
-      <ReduxStoreProvider>
-        <Layout>{children}</Layout>
-      </ReduxStoreProvider>
+      <ReduxStoreProvider>{children}</ReduxStoreProvider>
     </QueryProviders>
   );
 }
