+++ apps/frontend/src/app/layout.tsx
@@ -1,7 +1,8 @@
 import { Noto_Sans_KR } from "next/font/google";
 import "@repo/ui/styles/globals.css";
-import MSWProvider from "@/app/_providers/MSWProvider";
-import LayoutWithProviders from "./_providers/LayoutWithProviders";
+import MSWProvider from "@/_components/layout/providers/MSWProvider";
+import RootProviders from "@/_components/layout/providers/RootProviders";
+import LayoutUI from "@/_components/layout/LayoutUI";
 
 const noto_sans_kr = Noto_Sans_KR({
   subsets: ["latin"],
@@ -19,10 +20,14 @@ export default function RootLayout({
     <html lang="en">
       <body className={noto_sans_kr.className}>
         {!isUseMsw ? (
-          <LayoutWithProviders>{children}</LayoutWithProviders>
+          <RootProviders>
+            <LayoutUI>{children}</LayoutUI>
+          </RootProviders>
         ) : (
           <MSWProvider>
-            <LayoutWithProviders>{children}</LayoutWithProviders>
+            <RootProviders>
+              <LayoutUI>{children}</LayoutUI>
+            </RootProviders>
           </MSWProvider>
         )}
       </body>
